+++ lua/mini/colors.lua
@@ -431,8 +431,8 @@
 ---   - <general> `(boolean)` - general groups (like `Normal`). Default: `true`.
 ---   - <float> `(boolean)` - built-in groups for floating windows. Default: `false`.
 ---   - <statuscolumn> `(boolean)` - groups related to 'statuscolumn' (signcolumn,
----     numbercolumn, foldcolumn). Also updates groups for all currently
----     defined signs. Default: `false`.
+---     numbercolumn, foldcolumn, `DiagnosticSignXxx`, and `XxxMsg` groups). Also
+---     updates groups for all currently defined signs. Default: `false`.
 ---   - <statusline> `(boolean)` - built-in groups for 'statusline'. Default: `false`.
 ---   - <tabline> `(boolean)` - built-in groups for 'tabline'. Default: `false`.
 ---   - <winbar> `(boolean)` - built-in groups for 'winbar'. Default: `false`.
@@ -1297,12 +1297,15 @@ H.cs_add_transparency = function(self, opts)
 
   if opts.general then
     update({ 'Normal', 'NormalNC', 'EndOfBuffer', 'MsgArea', 'MsgSeparator', 'VertSplit', 'WinSeparator' })
+    update({ 'ErrorMsg', 'WarningMsg', 'OkMsg', 'ModeMsg', 'MoreMsg', 'StderrMsg', 'StdoutMsg' })
   end
 
   if opts.float then update({ 'FloatBorder', 'FloatTitle', 'NormalFloat' }) end
 
   if opts.statuscolumn then
     update({ 'FoldColumn', 'LineNr', 'LineNrAbove', 'LineNrBelow', 'SignColumn' })
+    local diag_hl = vim.tbl_map(function(x) return 'DiagnosticSign' .. x end, { 'Error', 'Warn', 'Info', 'Hint', 'Ok' })
+    update(diag_hl)
 
     -- Remove statuscolumn background coming from signs
     local signs = vim.fn.sign_getdefined()